Bugfixes - attempt to fix problems highlighted in https://github.com/tobixen/thrash-protect/issues/37
1) Sometimes a process may exit, but thrash-protect is still trying to
investigate, freeze or resume the said process.  There are some
try-except-blocks in the code for dealing with this, but not all
places were covered.

2) If the program exits due to an exception, it should try to clean up and
resume all processes.  This was implemented in 0.14.0, but never tested,
and apparently the code was buggy.

This commit has not been properly tested.  I will try to write up some test
code for both cases soon.

Sorry for the bad workflow practices.